core: Store different event types as a bitmask within the context
This change introduces a new event_flags member to the libusb_context
that holds a bitmask of different events that have occurred. This will
allow multiple "one-shot" events (those that don't require counting) to
be stored in a single variable. The only existing event of this type,
pollfds_modified, has been converted to use this bitmask instead.

Signed-off-by: Chris Dickens <christopher.a.dickens@gmail.com>